jobspy-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jobspy>=1.1.82
pandas>=2.1.0
pydantic>=2.3.0

